integrate linear token estimator directly into the main estimate_text_tokens function
