+++ tocount/__init__.py
@@ -2,6 +2,5 @@
 """Tocount modules."""
 
 from .params import TOCOUNT_VERSION
-from .heuristic.functions import estimate_text_tokens, TextEstimator
-from .linear import linear_tokens_estimator
+from .functions import estimate_text_tokens, TextEstimator
 __version__ = TOCOUNT_VERSION
