+++ tocount/linear/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 """Linear token estimator."""
 
-from ..params import LINEAR_MODELS
 from .estimator import linear_tokens_estimator
